backend/pyproject.toml
[project]
name = "apex-program"
version = "0.1.0"
description = "Life planner you need!"
readme = "README.md"
requires-python = ">=3.12,<=3.13"
dependencies = [
    "django>5.2,<5.3",
    "django-ckeditor>6,<7",
    "django-cors-headers>=4.7.0",
    "django-csp>=4.0",
    "django-filter>=25.1", # Filtering support
    "django-multi-captcha-admin>=2.0.0",
    "django-recaptcha==3.0.0",
    "django-robots>=6.1",
    "django-taggit>=6.1.0",
    "django-unfold>=0.75.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "jdatetime>=5.2.0",
    "markdown>=3.8.2", # Markdown support for the browsable API.
    "pillow>=12.1.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "pygraphviz>=1.14",  # for extensions
    "werkzeug>=3.1.3",  # for extensions
]
